fix bug, read data infinite loop

+++ coding/cacheserver/src/cacheserver.h
@@ -56,7 +56,7 @@ private:
 	bool m_bInnerRunOnce;
 	byte m_bytesInnerSrvBuffer[MsgBuffer::g_nReadBufferSize];
 	byte m_bytesInnerSrvOnceMsg[MsgBuffer::g_nOnceMsgSize];
-	ushort m_nHasReadProxyDataSize;
+	int m_nHasReadProxyDataSize;
 	// heart msg with proxy server
 	TimerCacheProxySrvHeart m_innerSrvHeart;
